requirements.txt
altgraph==0.17.4
appnope==0.1.4
asttokens==3.0.0
bottle==0.13.3
bottle-websocket==0.2.9
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Eel==0.18.1
executing==2.2.0
fonttools==4.56.0
fpdf2==2.8.2
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
greenlet==3.2.3
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
macholib==1.16.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.1
PySimpleGUI==5.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
pyzmq==26.2.1
requests==2.32.4
rsa==4.9.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.1
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.13
zope.event==5.0
zope.interface==7.2
